Bug fixes and movement improvements.
KatanaGrappleScript: Fixed a bug where the lineRenderer would be enabled while not grappling.
KatanaWeaponScript: Changed from a collision based attack to a raycast based attack.
WalkScript: Removed "Time.deltaTime" from the force calculations as the movement loop is "FixedUpdate()" which runs on a constant timer, no need to compensate for framerate fluctuations.