Setting up the enemy AI system.
The enemy AI runs on a state machine where each state is a part of the bahaviour of the enemy.
The state machine is contained inside a management script that sets up the states and their variables along with state transitions and the switching from one state to another.
This makes adding new states as easy as inherting from the StateInterface, writing the specific behaviour or or copying from an already existing state, wiring up the transitions and conditions and voila!