Added 2 more enemies and disabled VSync along with some name changes/typo corrections.
Added the BasicRanged enemy AI, which stays in a certain range away from the player (between backoffRange and approachRange), moves horizontally when in that range and then attacks.
Added the MobileBasicMelee enemy AI, it adds the ability for the enemy to dash left, right or forwards while approaching the player. May add some more behaviour later on.
disabled VSync for now, will probably add an option to toggle it.
Fixed some typos along with some class name changes (yes they're bad, I'll work on that).